python: update packaging setup, use pipenv, add CI steps
