+++ .github/workflows/ci.yml
@@ -9,7 +9,6 @@ on:
 jobs:
   typescript:
     runs-on: ubuntu-latest
-
     steps:
       - uses: actions/checkout@v2
       - uses: actions/setup-node@v2
@@ -20,3 +19,19 @@ jobs:
       - run: yarn install --frozen-lockfile
       - run: yarn workspace @foxglove/ws-protocol lint:ci
       - run: yarn workspace @foxglove/ws-protocol test
+
+  python:
+    runs-on: ubuntu-latest
+    defaults:
+      run:
+        working-directory: python
+    steps:
+      - uses: actions/checkout@v2
+      - uses: actions/setup-python@v2
+        with:
+          python-version: 3.8
+          cache: pipenv
+      - run: pipenv install --dev --deploy
+      - run: pipenv run black --check --diff --color
+      - run: pipenv run pyright src tests
+      - run: pipenv run pytest
