--- python/dev-requirements.in
@@ -1,2 +1,0 @@
--c requirements.txt
-black
