--- python/dev-requirements.txt
@@ -1,26 +1,0 @@
-#
-# This file is autogenerated by pip-compile with python 3.9
-# To update, run:
-#
-#    pip-compile dev-requirements.in
-#
-black==21.10b0
-    # via -r dev-requirements.in
-click==8.0.3
-    # via
-    #   -c requirements.txt
-    #   black
-mypy-extensions==0.4.3
-    # via black
-pathspec==0.9.0
-    # via black
-platformdirs==2.4.0
-    # via black
-regex==2021.11.10
-    # via black
-tomli==1.2.2
-    # via
-    #   -c requirements.txt
-    #   black
-typing-extensions==4.0.0
-    # via black
