--- python/requirements.in
@@ -1,3 +1,0 @@
-pip-tools
-websockets
-h5py
