--- python/requirements.txt
@@ -1,22 +1,0 @@
-#
-# This file is autogenerated by pip-compile with python 3.9
-# To update, run:
-#
-#    pip-compile
-#
-click==8.0.3
-    # via pip-tools
-pep517==0.12.0
-    # via pip-tools
-pip-tools==6.4.0
-    # via -r requirements.in
-tomli==1.2.2
-    # via pep517
-websockets==10.0
-    # via -r requirements.in
-wheel==0.37.0
-    # via pip-tools
-
-# The following packages are considered to be unsafe in a requirements file:
-# pip
-# setuptools
