+++ python/setup.cfg
@@ -1,0 +1,22 @@
+[metadata]
+name = foxglove-websocket
+version = 0.0.1
+description = Foxglove Studio WebSocket server
+url = https://github.com/foxglove/ws-protocol
+project_urls =
+    Bug Tracker = https://github.com/foxglove/ws-protocol/issues
+classifiers =
+    Programming Language :: Python :: 3
+    License :: OSI Approved :: MIT License
+    Operating System :: OS Independent
+
+[options]
+package_dir =
+    = src
+packages = find:
+python_requires = >=3.8
+install_requires =
+    websockets
+
+[options.packages.find]
+where = src
